+++ qcelemental/models/__init__.py
@@ -1,1 +1,10 @@
+try:
+    import pydantic
+except ImportError:  # pragma: no cover
+    raise ImportError(
+        "Python module pydantic not found. Solve by installing it: "
+        "`conda install pydantic -c conda-forge` or `pip install pydantic`"
+    )
+
+from . import v1, v2
 from .v1 import *