+++ qcelemental/models/v2/common_models.py
@@ -1,18 +1,18 @@
 from enum import Enum
-from typing import TYPE_CHECKING, Any, Dict, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, Optional, Sequence, Tuple, Union
 
 import numpy as np
-from pydantic.v1 import Field
+from pydantic import Field
 
 from .basemodels import ProtoModel, qcschema_draft
 from .basis import BasisSet
 
 if TYPE_CHECKING:
-    from pydantic.v1.typing import ReprArgs
+    ReprArgs = Sequence[Tuple[Optional[str], Any]]
 
 
-# Encoders, to be deprecated
-ndarray_encoder = {np.ndarray: lambda v: v.flatten().tolist()}
+def provenance_json_schema_extra(schema, model):
+    schema["$schema"] = qcschema_draft
 
 
 class Provenance(ProtoModel):
@@ -21,16 +21,14 @@ class Provenance(ProtoModel):
     creator: str = Field(..., description="The name of the program, library, or person who created the object.")
     version: str = Field(
         "",
-        description="The version of the creator, blank otherwise. This should be sortable by the very broad `PEP 440 <https://www.python.org/dev/peps/pep-0440/>`_.",
+        description="The version of the creator, blank otherwise. "
+        "This should be sortable by the very broad `PEP 440 <https://www.python.org/dev/peps/pep-0440/>`_.",
     )
     routine: str = Field("", description="The name of the routine or function within the creator, blank otherwise.")
 
-    class Config(ProtoModel.Config):
-        canonical_repr = True
-        extra: str = "allow"
-
-        def schema_extra(schema, model):
-            schema["$schema"] = qcschema_draft
+    model_config = ProtoModel._merge_config_with(
+        canonical_repr=True, json_schema_extra=provenance_json_schema_extra, extra="allow"
+    )
 
 
 class Model(ProtoModel):
@@ -48,10 +46,7 @@ class Model(ProtoModel):
     )
 
     # basis_spec: BasisSpec = None  # This should be exclusive with basis, but for now will be omitted
-
-    class Config(ProtoModel.Config):
-        canonical_repr = True
-        extra: str = "allow"
+    model_config = ProtoModel._merge_config_with(canonical_repr=True, extra="allow")
 
 
 class DriverEnum(str, Enum):
@@ -75,7 +70,8 @@ class ComputeError(ProtoModel):
 
     error_type: str = Field(  # type: ignore
         ...,  # Error enumeration not yet strict
-        description="The type of error which was thrown. Restrict this field to short classifiers e.g. 'input_error'. Suggested classifiers: https://github.com/MolSSI/QCEngine/blob/master/qcengine/exceptions.py",
+        description="The type of error which was thrown. Restrict this field to short classifiers e.g. 'input_error'. "
+        "Suggested classifiers: https://github.com/MolSSI/QCEngine/blob/master/qcengine/exceptions.py",
     )
     error_message: str = Field(  # type: ignore
         ...,
@@ -87,15 +83,17 @@ class ComputeError(ProtoModel):
         description="Additional information to bundle with the error.",
     )
 
-    class Config:
-        repr_style = ["error_type", "error_message"]
+    model_config = ProtoModel._merge_config_with(repr_style=["error_type", "error_message"])
 
     def __repr_args__(self) -> "ReprArgs":
         return [("error_type", self.error_type), ("error_message", self.error_message)]
 
 
 class FailedOperation(ProtoModel):
-    """Record indicating that a given operation (program, procedure, etc.) has failed and containing the reason and input data which generated the failure."""
+    """
+    Record indicating that a given operation (program, procedure, etc.) has failed
+    and containing the reason and input data which generated the failure.
+    """
 
     id: str = Field(  # type: ignore
         None,